+++ apps/examples/blinkid-advanced-setup/CHANGELOG.md
@@ -1,5 +1,15 @@
 # example-blinkid-advanced-setup
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [5d08204]
+- Updated dependencies [5d08204]
+  - @microblink/blinkid-ux-manager@7.2.0
+  - @microblink/feedback-stabilizer@7.1.1
+  - @microblink/blinkid-core@7.2.0
+
 ## 1.1.3
 
 ### Patch Changes