+++ packages/blinkid-worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microblink/blinkid-worker
 
+## 7.2.0
+
+### Patch Changes
+
+- Updated dependencies [3655cc6]
+- Updated dependencies [5d08204]
+  - @microblink/blinkid-wasm@7.2.0
+
 ## 7.1.0
 
 ### Patch Changes