+++ packages/feedback-stabilizer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @microblink/feedback-stabilizer
 
+## 7.1.1
+
+### Patch Changes
+
+- 5d08204: Stabilizer properly waits for previous state to finish on success events
+
 ## 7.1.0
 
 ### Minor Changes